+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.8.3] - 2022-04-25
+
+### Fixed
+
+- Calculations that were incorrect for multiple modes using `FockSimulator`
+  along with `Squeezing`, `Displacement` or `CubicPhase`.
+
 ## [0.8.2] - 2022-04-07
 
 ### Fixed